ktr_platform/flows/3aaa9a2e-831f-4a0a-ace2-51ebd88baf7c/requirements.txt
loguru>=0.7.0
openpyxl>=3.1.0
pandas>=2.0.0
python-decouple>=3.8
sqlalchemy>=2.0.0